large project structure (v0.10)
