+++ orders/app/exceptions.py
@@ -1,0 +1,4 @@
+class ValidationError(ValueError):
+    pass
+
+
