+++ orders/app/utils.py
@@ -1,6 +1,7 @@
 from flask.globals import _app_ctx_stack, _request_ctx_stack
 from werkzeug.urls import url_parse
 from werkzeug.exceptions import NotFound
+from .exceptions import ValidationError
 
 
 def split_url(url, method='GET'):
