+++ orders/tests/__init__.py
@@ -1,0 +1,4 @@
+import unittest
+from .tests import TestAPI
+
+suite = unittest.TestLoader().loadTestsFromTestCase(TestAPI)
